refactor(sessions): move pipeline_watch into a package, unchanged
