+++ daiv/sessions/pipeline_watch/__init__.py
@@ -1,0 +1,2 @@
+"""CI watch on a merge request DAIV published: judge the pipeline, dispatch a bounded number of
+fix runs, and give up loudly when it cannot make progress."""
